requirements.txt
revChatGPT
flask
peewee
psycopg2
python-dotenv
requests
python-slugify
markdown
